pyproject.toml
[tool.poetry]
name = "reproscreener"
version = "0.1.0"
description = "Automated machine learning model verification"
authors = ["Adhithya Bhaskar <adhi@adhi.dev>"]
license = "MIT"
packages = [{include = "reproscreener", from = "src"}]

[tool.poetry.scripts]
reproscreener = "reproscreener.main:app"

[tool.poetry.dependencies]
python = "=3.9.13"
arxivscraper = "^0.0.5"
requests = "^2.30.0"
pandas = "^2.0.1"
pathlib = "^1.0.1"
flashtext = "^2.7"
exrex = "^0.11.0"
lxml = "^4.9.2"
rich = "^13.3.5"
typer = "^0.9.0"
gitpython = "^3.1.31"
pydrive2 = "^1.15.3"
urlextract = "^1.8.0"
uritools = "^4.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = {extras = ["jupyter"], version = "^23.3.0"}
mypy = "^1.3.0"
loguru = "^0.7.0"
beautifulsoup4 = "^4.12.2"
bandit = "^1.7.5"
seaborn = "^0.12.2"
types-requests = "^2.30.0.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pylint = "^2.17.4"
vulture = "^2.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
